Add missing checks for the image size for svg (Denis Robinet)
